// practitioner/src/app/models/tour.ts
export enum TourType {
    HOME = 'home',
    DASHBOARD = 'dashboard',
    AVAILABILITY = 'availability',
    WAITING_ROOM = 'waiting-room',
    OPEN_CONSULTATIONS = 'open-consultations',
    CLOSED_CONSULTATIONS = 'closed-consultations',
    
    INVITES_MENU = 'invites',
    WAITING_ROOM_MENU = 'waiting-room',
    OPENED_CONSULTATIONS_MENU = 'open-consultations',
    CONSULTATION_HISTORY_MENU = 'closed-consultations',
    HEADER_PROFILE_MENU = 'admin-logo-wrapper',
    
    DASHBOARD_INVITE_BUTTON = 'dashboard-invite-button',
    VIEW_ALL_BUTTON = 'view-all-button',
    HELP_BUTTON = 'help-button',
    
    INVITE_FORM_PATIENT_INFO = 'invite-form-patient-info',
    INVITE_FORM_GUEST_OPTIONS = 'invite-form-guest-options',
    INVITE_FORM_MANUAL_SEND = 'invite-form-manual-send',
    INVITE_FORM_SUBMIT = 'invite-form-submit',
    
    OPEN_CONSULTATIONS_PAGE = 'open-consultations-page',
    OPEN_CONSULTATIONS_GRID = 'open-consultations-grid',
    OPEN_CONSULTATION_PANEL = 'app-open-consultation-panel',
    CONSULTATION_HISTORY_PAGE = 'consultation-history-page',
    CONSULTATION_HISTORY_GRID = 'consultation-history-grid',
    
    AVAILABILITY_FORM_CARD = 'availability-form-card',
    AVAILABILITY_TABLE = 'availability-table',
    AVAILABILITY_FORM = 'availability-form',
    AVAILABILITY_TABLE_CARD = 'availability-table-card',
    
    PROFILE_PERSONAL_INFO = 'profile-personal-info',
    PROFILE_PROFESSIONAL_INFO = 'profile-professional-info'
}